caption-it/requirements.txt
Flask==2.1.2
gunicorn==20.1.0
libclang==14.0.1

keras==2.9.0
Keras-Preprocessing==1.1.2
matplotlib-inline==0.1.3
nltk==3.7
notebook==6.4.11
numpy==1.22.4
pandas==1.4.2
scikit-learn==1.1.1
tensorflow-cpu==2.9.1
